Cleaned envs Removed old contributing Updated readme.rst
