--- learnrl/environments/single_agents/random_crosses_and_noughts/__init__.py
@@ -1,5 +1,0 @@
-# LearnRL a python library to learn and use reinforcement learning
-# Copyright (C) 2020 Mathïs FEDERICO <https://www.gnu.org/licenses/>
-
-from learnrl.environments.single_agents.random_crosses_and_noughts.game import RdCrossesAndNoughtsEnv
-
